requirements.txt
appdirs==1.4.3
feedparser==5.2.1
google-api-python-client==1.6.2
httplib2==0.10.3
oauth2client==4.1.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.14.2
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
